requirements.txt
fastapi
pydantic
docker

fastapi
uvicorn[standard]
passlib[bcrypt]
pyjwt
python-multipart

httpx
pytest

beanie